✨ Geocoding class with custom collection + item classes
The Geocoding class is done, it is now possible to retrieve the geocoding results for coordinates or an address using a fluent API.

The results returned are stored in cache for the duration set in the config file.

The results returned are defined as a collection (customizable in the config) which contains an array of classes (customizable in the config) which can be returned as array.

When returned as an array, the collection internally calls the toArray method on each element.

This method can be customized using a different item class via the config, by default is will return an array containing the latitude, longitude, street name, street number, country, region, city and postal code of the result.