✨ Remove components when calling the coordinates method by default
The latlng geocoding process doesn't work well with the components (it usually just returns the content of the component, for example if country is set to BE is will return an address_components containing the country Belgium, if the postal_code component is set to 4020 it will return multiple address_components containing the city, postal code, region and country of said postal code... So nothing related to the coordinates).

If the user still wants to keep the components in the url he can set the third parameter to true like this: coordinates($lat, $lng, true) or use the coordinatesWithComponents($lat, $lng).